implementing file backend

+++ src/endpoints/contexts/file/LocalFsFilePersistenceProvider.ts
@@ -3,14 +3,16 @@ import {noopAsync} from '../../../utils/fns';
 import {
   FilePersistenceDeleteFilesParams,
   FilePersistenceGetFileParams,
-  FilePersistenceProviderContext,
+  FilePersistenceProvider,
   FilePersistenceUploadFileParams,
-  IPersistedFile,
+  PersistedFile,
 } from './types';
 
-export default class LocalFsFilePersistenceProviderContext
-  implements FilePersistenceProviderContext
-{
+/**
+ * - Does not use the `bucket` param, so it's okay passing an empty string for
+ *   that
+ */
+export default class LocalFsFilePersistenceProvider implements FilePersistenceProvider {
   constructor(private fileDir: string) {
     fse.ensureDirSync(this.fileDir);
   }
@@ -19,14 +21,17 @@ export default class LocalFsFilePersistenceProviderContext
     const filepath = `${this.fileDir}/${params.key}`;
     await fse.ensureFile(filepath);
     return new Promise<void>((resolve, reject) => {
-      const writeStream = fse.createWriteStream(filepath, {autoClose: true, emitClose: true});
+      const writeStream = fse.createWriteStream(filepath, {
+        autoClose: true,
+        emitClose: true,
+      });
       writeStream.on('close', resolve);
       writeStream.on('error', reject);
       params.body.pipe(writeStream);
     });
   };
 
-  getFile = async (params: FilePersistenceGetFileParams): Promise<IPersistedFile> => {
+  getFile = async (params: FilePersistenceGetFileParams): Promise<PersistedFile> => {
     const filepath = `${this.fileDir}/${params.key}`;
     const stat = await fse.promises.stat(filepath);
 
@@ -39,13 +44,18 @@ export default class LocalFsFilePersistenceProviderContext
   };
 
   deleteFiles = async (params: FilePersistenceDeleteFilesParams) => {
+    if (params.keys.length === 0) {
+      // Short-circuit, no files to delete
+      return;
+    }
+
     const rmPromises = params.keys.map(async key => {
       const filepath = `${this.fileDir}/${key}`;
       await fse.promises.rm(filepath);
     });
+
     await Promise.all(rmPromises);
   };
 
-  ensureBucketReady = noopAsync;
   close = noopAsync;
 }
